fix: Move NewsColumns definition into a custom hook
- Refactored the NewsColumns definition to be inside a custom hook (useNewsColumns) to comply with React's hooks rules.
- Created a NewsTable component to utilize the memoized columns for rendering.
- Improved code organization and maintainability.
