+++ src/app/dashboard/news/components/columns.jsx
@@ -1,40 +1,45 @@
 import Image from "next/image";
-import RowActions from "./rowActions";
+import { useMemo } from "react";
+import NewsRowActions from "./NewsRowAction";
 
-export const columns = [
-    {
+export const useNewsColumns = () => {
+  return useMemo(() => {
+    return [
+      {
         accessorKey: "images",
         header: "Image",
         cell: ({ row }) => {
-            const image = row.getValue("images");
+          const image = row.getValue("images");
 
-            return (
-                <div>
-                    <Image
-                        src={image}
-                        alt="Image"
-                        width={80}
-                        height={60}
-                        className="bg-gray-100 rounded-md"
-                    />
-                </div>
-            );
+          return (
+            <div>
+              <Image
+                src={image}
+                alt="Image"
+                width={80}
+                height={60}
+                className="bg-gray-100 rounded-md"
+              />
+            </div>
+          );
         },
-    },
-    {
+      },
+      {
         accessorKey: "title",
         header: "News title",
-    },
-    {
+      },
+      {
         accessorKey: "newsCategory",
         header: "Category",
-    },
-    {
+      },
+      {
         id: "actions",
         cell: ({ row }) => {
-            const news = row.original;
+          const news = row.original;
 
-            return <RowActions news={news} />;
+          return <NewsRowActions news={news} />;
         },
-    },
-];
+      },
+    ];
+  }, []);
+};
